Change Battery capacity type to double add checkLoadedMedication in MedicationService Handle BusinessException add method findByName to MedicationRepository.jar remove Drone MedicationLoad add List of Medication to Drone

+++ src/main/java/com/musalasoft/dronesServiceDelivering/service/BatteryService.java
@@ -1,12 +1,10 @@
 package com.musalasoft.dronesServiceDelivering.service;
 
-import java.math.BigDecimal;
-
 /**
  * @author Ahmed.Yusef
  *
  * @Date Apr 15, 2022 5:49:02 PM
  */
 public interface BatteryService {
-	BigDecimal checkBatteryLevel(String serialNumber);
+	double checkBatteryLevel(String serialNumber);
 }
